Fix KafkaContainer not running as expected on ARM64
There are two issues with the current implementation.

The first is that the ARM64 image (`niciqy/cp-kafka-arm64`) is not
marked as a compatible substitute for the confluent image
(`confluentinc/cp-kafka`). This means a runtime failure executing tests.

The second is that `System.getProperty("os.arch")` does not return
`aarch64` on an x86 (AMD64) JVM being emulated by Rosetta 2 on MacOS
(for example because other parts of compilation/testing require an x86
JDK). End-users _probably_ want to use an ARM64 Kafka image in this
situation as they (probably) won't need an emulated Kafka container
running. A call to `sysctl` can determine if the current process (the
JVM) is running under Rosetta 2 and so if the underlying CPU is actually
ARM64.

These changes are being used internally and have worked (so far!) for
users on both x86 and ARM64 MacOS machines.